implement export for oas30, refactor
