+++ localstack/services/apigateway/helpers.py
@@ -7,10 +7,9 @@ import re
 import time
 from collections import defaultdict
 from datetime import datetime, timezone
-from typing import Any, Callable, Dict, List, Optional, Tuple, TypedDict, Union
+from typing import Any, Dict, List, Optional, Tuple, TypedDict, Union
 from urllib import parse as urlparse
 
-from apispec import APISpec
 from botocore.utils import InvalidArnException
 from jsonpatch import apply_patch
 from jsonpointer import JsonPointerException
@@ -27,7 +26,6 @@ from localstack.aws.api.apigateway import (
     DocumentationPart,
     DocumentationPartLocation,
     IntegrationType,
-    ListOfModel,
     Model,
     RequestValidator,
 )
@@ -52,7 +50,6 @@ from localstack.utils.aws.aws_responses import requests_error_response_json, req
 from localstack.utils.aws.request_context import MARKER_APIGW_REQUEST_REGION, THREAD_LOCAL
 from localstack.utils.json import try_json
 from localstack.utils.strings import long_uid, short_uid, to_bytes, to_str
-from localstack.utils.time import TIMESTAMP_FORMAT_TZ, timestamp
 from localstack.utils.urls import localstack_host
 
 LOG = logging.getLogger(__name__)
@@ -1496,165 +1493,6 @@ def create_invocation_headers(invocation_context: ApiInvocationContext) -> Dict[
             ):
                 headers.update({header_name: req_parameter_value})
     return headers
-
-
-# TODO:
-# - handle extensions
-#
-
-TypeExporter = Callable[[str, str, str], str]
-
-
-class OpenApiExporter:
-    SWAGGER_VERSION = "2.0"
-    OPENAPI_VERSION = "3.0.1"
-
-    exporters: Dict[str, TypeExporter]
-
-    def __init__(self):
-        self.exporters = {"swagger": self._swagger_export, "oas30": self._oas30_export}
-        self.export_formats = {"application/json": "to_dict", "application/yaml": "to_yaml"}
-
-    def export_api(
-        self, api_id: str, stage: str, export_type: str, export_format: str = "application/json"
-    ) -> str:
-        return self.exporters.get(export_type)(api_id, stage, export_format)
-
-    @staticmethod
-    def _add_paths(spec, resources):
-        for item in resources.get("items"):
-            path = item.get("path")
-            for method, method_config in item.get("resourceMethods", {}).items():
-                print(f"{path=}, {method_config=}")
-                method = method.lower()
-
-                # TODO: this is extension, integration is extensions
-                method_integration = method_config.get("methodIntegration", {})
-                integration_responses = method_integration.get("integrationResponses", {})
-                method_responses = method_config.get("methodResponses")
-                responses = {}
-                produces = set()
-                for status_code, values in method_responses.items():
-                    response = {"description": f"{status_code} response"}
-                    if response_parameters := values.get("responseParameters"):
-                        headers = {}
-                        for parameter in response_parameters:
-                            in_, name = parameter.removeprefix("method.response.").split(".")
-                            # TODO: other type?
-                            if in_ == "header":
-                                headers[name] = {"type": "string"}
-
-                        if headers:
-                            response["headers"] = headers
-                    if response_models := values.get("responseModels"):
-                        for content_type, model_name in response_models.items():
-                            produces.add(content_type)
-                            response["schema"] = model_name
-                    if integration_response := integration_responses.get(status_code, {}):
-                        produces.update(integration_response.get("responseTemplates", {}).keys())
-
-                    responses[status_code] = response
-
-                request_parameters = method_config.get("requestParameters", {})
-                parameters = []
-                for parameter, required in request_parameters.items():
-                    in_, name = parameter.removeprefix("method.request.").split(".")
-                    in_ = in_ if in_ != "querystring" else "query"
-                    parameters.append(
-                        {"name": name, "in": in_, "required": required, "type": "string"}
-                    )
-
-                request_models = method_config.get("requestModels", {})
-                for model_name in request_models.values():
-                    parameter = {
-                        "in": "body",
-                        "name": model_name,
-                        "required": True,
-                        "schema": {"$ref": f"#/definitions/{model_name}"},
-                    }
-                    parameters.append(parameter)
-
-                content_types = request_models | method_integration.get("requestTemplates", {})
-
-                method_operations = {"responses": responses}
-                if parameters:
-                    method_operations["parameters"] = parameters
-                if produces:
-                    method_operations["produces"] = list(produces)
-                if content_types:
-                    method_operations["consumes"] = list(content_types.keys())
-                if operation_name := method_config.get("operationName"):
-                    method_operations["operationId"] = operation_name
-
-                spec.path(path=path, operations={method: method_operations})
-
-    def _get_definitions(self, models: ListOfModel):
-        definitions = {}
-        for model in models:
-            model_def = json.loads(model["schema"])
-            self._resolve_refs(model_def)
-            definitions[model["name"]] = model_def
-        return definitions
-
-    def _resolve_refs(self, schema: dict):
-        if "$ref" in schema:
-            schema["$ref"] = f'#/definitions/{schema["$ref"].rsplit("/", maxsplit=1)[-1]}'
-        for value in schema.values():
-            if isinstance(value, dict):
-                self._resolve_refs(value)
-
-    def _swagger_export(self, api_id: str, stage: str, export_format: str) -> str:
-        """
-        https://github.com/OAI/OpenAPI-Specification/blob/main/versions/2.0.md
-        """
-        apigateway_client = connect_to().apigateway
-
-        rest_api = apigateway_client.get_rest_api(restApiId=api_id)
-        resources = apigateway_client.get_resources(restApiId=api_id)
-        models = apigateway_client.get_models(restApiId=api_id)
-        definitions = self._get_definitions(models["items"])
-
-        info = {}
-        kwargs = {}
-        if (description := rest_api.get("description")) is not None:
-            info["description"] = description
-        if definitions:
-            kwargs["definitions"] = definitions
-
-        spec = APISpec(
-            title=rest_api.get("name"),
-            version=rest_api.get("version")
-            or timestamp(rest_api.get("createdDate"), format=TIMESTAMP_FORMAT_TZ),
-            info=info,
-            openapi_version=self.SWAGGER_VERSION,
-            basePath=f"/{stage}",
-            schemes=["https"],
-            **kwargs,
-        )
-
-        self._add_paths(spec, resources)
-
-        return getattr(spec, self.export_formats.get(export_format))()
-
-    def _oas30_export(self, api_id: str, stage: str, export_format: str) -> str:
-        """
-        https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md
-        """
-        apigateway_client = connect_to().apigateway
-
-        rest_api = apigateway_client.get_rest_api(restApiId=api_id)
-        resources = apigateway_client.get_resources(restApiId=api_id)
-
-        spec = APISpec(
-            title=rest_api.get("name"),
-            version=timestamp(rest_api.get("createdDate"), format=TIMESTAMP_FORMAT_TZ),
-            info=dict(description=rest_api.get("description")),
-            openapi_version=self.OPENAPI_VERSION,
-        )
-
-        self._add_paths(spec, resources)
-
-        return getattr(spec, self.export_formats.get(export_format))()
 
 
 def is_greedy_path(path_part: str) -> bool:
